misc/HtmlTestSuite/results/test-0911.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>document.close called while a script is pending</title>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
</head><body>
  <script>
    window.t = async_test();
    // We want start a document load, create an non-blocking script load inside
    // it, have the parser complete, then call document.open()/document.close()
    // after the parser is done but before the non-blocking script load
    // completes.  After we do that, the document should reach the 'complete'
    // ready state and the iframe's load event should fire.
    var loadFired = false;
    var i;

    var finish = t.step_func_done(() =&gt; {
      assert_equals(loadFired, true, "Should have fired a load event");
      assert_equals(i.contentDocument.readyState, "complete",
                    "Should be fully loaded");
    });

    var checkForLoad = t.step_func(() =&gt; {
      if (loadFired) {
        finish();
      } else {
        i.addEventListener("load", finish);
      }
    });

    window.parserDone = t.step_func(() =&gt; {
      var doc = i.contentDocument;
      i.onload = () =&gt; { loadFired = true; }
      doc.open();
      doc.close();
      // It's not very clearly specced whether the document is
      // supposed to be fully loaded at this point or not, so allow
      // that to be the case, or to happen soonish.
      assert_true(doc.readyState == "interactive" ||
                  doc.readyState == "complete", "Should be almost loaded");
      if (doc.readyState == "complete") {
        checkForLoad();
      } else {
        doc.addEventListener("readystatechange", checkForLoad);
      }
    });

    t.step(() =&gt; {
        i = document.createElement("iframe");
        i.srcdoc = `
          &lt;script&gt;
            parent.t.step(() =&gt; {
              var s = document.createElement("script");
              s.src = "/common/slow.py";
              document.documentElement.appendChild(s);
              // Call into the parent async, so we finish our "end of parse"
              // work before it runs.
              document.addEventListener(
                "DOMContentLoaded",
                () =&gt; parent.t.step_timeout(parent.parserDone, 0));
            });
          &lt;\/script&gt;
        `;
        document.body.appendChild(i);
    });
  </script>
</body>
</html>
